+++ src/lightning/pytorch/core/module.py
@@ -21,7 +21,6 @@ from pathlib import Path
 from typing import (
     Any,
     Callable,
-    cast,
     Dict,
     Generator,
     IO,
@@ -1447,7 +1446,7 @@ class LightningModule(
         hparams_file: Optional[_PATH] = None,
         strict: bool = True,
         **kwargs: Any,
-    ) -> Self:
+    ) -> Self:  # type: ignore[valid-type]
         r"""
         Primary way of loading a model from a checkpoint. When Lightning saves a checkpoint
         it stores the arguments passed to ``__init__``  in the checkpoint under ``"hyper_parameters"``.
@@ -1519,7 +1518,7 @@ class LightningModule(
             pretrained_model.freeze()
             y_hat = pretrained_model(x)
         """
-        loaded = _load_from_checkpoint(
+        return _load_from_checkpoint(
             cls,
             checkpoint_path,
             map_location,
@@ -1527,7 +1526,6 @@ class LightningModule(
             strict,
             **kwargs,
         )
-        return cast(Self, loaded)
 
     @contextmanager
     def _prevent_trainer_and_dataloaders_deepcopy(self) -> Generator[None, None, None]: